TzotzilBibleApp-v3/requirements.txt

flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.0.5
openai==1.3.5
python-dotenv==1.0.0
faiss-cpu==1.7.4
numpy==1.24.3
spacy==3.7.2
transformers==4.35.2
python-jose==3.3.0
cachetools==5.3.1
